# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEC",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/",
            "description": "DEV"
        }
    ],
    "paths": {
...
        "/groups": {
            "get": {
                "tags": [
                    "groups"
                ],
                "summary": "Listar todos os grupos",
                "description": "Lista todos os grupos",
                "operationId": "bef7bc6436ad2524e81253beb2657bf9",
                "parameters": [
                    {
                        "name": "entity",
                        "in": "query",
                        "description": "nome da entidade que deseja filtrar",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "groups"
                ],
                "summary": "Criar novo grupo",
                "description": "Cria novo grupo",
                "operationId": "1066d2dc9a3c9d46e0b28da7505b4ac9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "entity": "exercitationem",
                                    "organ": "architecto",
                                    "council": "voluptates",
                                    "acronym": "nesciunt",
                                    "team": "ut",
                                    "unit": "occaecati",
                                    "email": "amanda24@hotmail.com",
                                    "office_requested": "accusamus",
                                    "office_indicated": "incidunt",
                                    "internal_concierge": "corrupti",
                                    "type_group_id": 1,
                                    "observations": "Repellendus aut voluptatem quaerat consequuntur illum. Dolor est sed natus est. Qui voluptatibus iure necessitatibus velit.",
                                    "status": "EM ANDAMENTO",
                                    "representative": "representante@mail.com",
                                    "name": "comissão",
                                    "type_group": "interno"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Apenas usuarios do tipo representante sao permitidos"
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "tags": [
                    "groups"
                ],
                "summary": "Lista o registro de grupos por ID",
                "description": "Lista o registro de grupos por ID de referência",
                "operationId": "76a20ea28a20c0879a3588fd099f5651",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do grupo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "grupo not found"
                    },
                    "200": {
                        "description": "Ok"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "groups"
                ],
                "summary": "Atualizar grupo",
                "description": "Atualizar grupo",
                "operationId": "0841409d1cb106baef2ac166a7bda69b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do grupo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "entity": "exercitationem",
                                    "organ": "architecto",
                                    "council": "voluptates",
                                    "acronym": "nesciunt",
                                    "team": "ut",
                                    "unit": "occaecati",
                                    "email": "amanda24@hotmail.com",
                                    "office_requested": "accusamus",
                                    "office_indicated": "incidunt",
                                    "internal_concierge": "corrupti",
                                    "name": "comissão",
                                    "type_group": "interno",
                                    "observations": "Repellendus aut voluptatem quaerat consequuntur illum. Dolor est sed natus est. Qui voluptatibus iure necessitatibus velit.",
                                    "status": "EM ANDAMENTO",
                                    "representative": "representante@mail.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Usuário not found"
                    },
                    "400": {
                        "description": "Apenas usuarios do tipo representante sao permitidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "groups"
                ],
                "summary": "Deletar grupo",
                "description": "Deletargrupo por ID de referência",
                "operationId": "034d1a4e1246265f300642aed8b0b455",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do grupo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "grupo Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ],
...
}